Fixed bug that Xavier noticed. Count is a Long, not an Integer
Former-commit-id: a0d229915d54747bf03d558471050838f0511545

+++ intermine/objectstore/main/src/org/intermine/objectstore/query/QueryFunction.java
@@ -90,7 +90,7 @@ public class QueryFunction implements QueryEvaluable
        */
     public Class getType() {
         if (op == COUNT) {
-            return Integer.class;
+            return Long.class;
         }
         return obj.getType();
     }
